feat: add logic to deny delete user if they have tasks related

+++ src/main/java/com/weg/rocketseatcourse/application/usecase/user/implementation/DeleteUserUseCaseImpl.java
@@ -1,9 +1,13 @@
 package com.weg.rocketseatcourse.application.usecase.user.implementation;
 
+import com.weg.rocketseatcourse.application.exceptions.UserCantBeDeleted;
 import com.weg.rocketseatcourse.application.exceptions.UserNotFoundException;
 import com.weg.rocketseatcourse.application.usecase.user.interfaces.DeleteUserUseCase;
+import com.weg.rocketseatcourse.domain.entity.Task;
 import com.weg.rocketseatcourse.domain.entity.User;
+import com.weg.rocketseatcourse.domain.repository.TaskRepository;
 import com.weg.rocketseatcourse.domain.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Component;
 
 import java.util.UUID;
@@ -12,15 +16,22 @@ import java.util.UUID;
 public class DeleteUserUseCaseImpl implements DeleteUserUseCase {
 
     private final UserRepository userRepository;
+    private final TaskRepository taskRepository;
 
-    public DeleteUserUseCaseImpl (UserRepository userRepository){
+    public DeleteUserUseCaseImpl (UserRepository userRepository, TaskRepository taskRepository){
         this.userRepository = userRepository;
+        this.taskRepository = taskRepository;
     }
 
+    @Transactional
     @Override
     public void deleteUserById(UUID id) {
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException("User with this Id not found!"));
+
+        if(taskRepository.existsByUserId(user.getId())){
+            throw new UserCantBeDeleted("You can't delete a user with tasks related!");
+        }
 
         userRepository.deleteById(user.getId());
     }
